Refact: .h -> .hh
```sh
for f in include/nova/*; do echo -e "// Deprecated. Proxy header for including '.hh' files.\n#pragma once\n\n#include \"nova/${f##*/}\"" > "${f%%.*}".h; done
for f in include/nova/*.hh; do sed -i 's/\(#include "nova.*\).h"/\1.hh"/' "$f"; done
```
