+++ include/nova/main.h
@@ -1,29 +1,4 @@
-/**
- * Part of Nova C++ Library.
- *
- * The modern main.
- */
-
+// Deprecated. Proxy header for including '.hh' files.
 #pragma once
 
-#include <spdlog/spdlog.h>
-
-#include <ranges>
-#include <string_view>
-#include <span>
-
-#define MAIN(func)                                                              \
-    int main(int argc, char* argv[]) {                                          \
-        try {                                                                   \
-            const auto args = std::span(argv, static_cast<std::size_t>(argc));  \
-            func(args | std::views::transform(                                  \
-                [](const auto& x) { return std::string_view(x); })              \
-            );                                                                  \
-        } catch (std::exception& ex) {                                          \
-            spdlog::error("Exception caught in main: {}", ex.what());           \
-        } catch (const char* msg) {                                             \
-            spdlog::error("Exception caught in main: {}", msg);                 \
-        } catch (...) {                                                         \
-            spdlog::error("Unknown exception caught in main");                  \
-        }                                                                       \
-    }
+#include "nova/main.hh"
