+++ include/nova/nova.h
@@ -1,35 +1,4 @@
-/**
- * Part of Nova C++ Library.
- *
- * Main wrapper header.
- *
- * Contains statically known information about the library.
- */
-
+// Deprecated. Proxy header for including '.hh' files.
 #pragma once
 
-#include <nova/color.h>
-#include <nova/data.h>
-#include <nova/error.h>
-#include <nova/expected.h>
-#include <nova/flat_map.h>
-#include <nova/intrinsics.h>
-#include <nova/io.h>
-#include <nova/json.h>
-#include <nova/main.h>
-#include <nova/nova.h>
-#include <nova/random.h>
-#include <nova/static_string.h>
-#include <nova/std_extensions.h>
-#include <nova/system.h>
-#include <nova/threading.h>
-#include <nova/type_traits.h>
-#include <nova/types.h>
-#include <nova/units.h>
-#include <nova/utils.h>
-#include <nova/vec.h>
-#include <nova/yaml.h>
-
-constexpr auto NovaVersionMajor = 0;
-constexpr auto NovaVersionMinor = 1;
-constexpr auto NovaVersionPatch = 0;
+#include "nova/nova.hh"
