+++ include/nova/types.hh
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace nova {
+
+template <typename T>
+struct range {
+    T low;
+    T high;
+};
+
+} // namespace nova
